completed all view pages

+++ Allprojects/app/Http/Controllers/web/NoteController.php
@@ -6,16 +6,21 @@ use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use App\Http\Controllers\API\BaseController as BaseController;
 use App\Http\Resources\Note as NoteResource;
 
 
-class NoteController extends Controller
+class NoteController extends BaseController
 {
+	public function __construct()
+	{
+		$this->middleware('auth');
+	}
     
     public function userNotes()
     {
-    	$note= Note::all();
-    	return view('notes.userNotes')->with('note', $note);
+    	$notes= Note::all();
+    	return view('notes.userNotes')->with('notes', $notes);
     }
 
      public function create()
@@ -32,6 +37,11 @@ class NoteController extends Controller
             'content'=>'required'
         ]);
 
+         if ($validator->fails()) {
+            return $this->sendErorr('Validate Error',$validator->errors());
+        }
+
+
        $user=Auth::user();
        $input['user_id']=$user->id;
        $note=Note::create($input);
@@ -44,6 +54,9 @@ class NoteController extends Controller
      public function show($id)
     {
     	$note=Note::find($id);
+    	if ( $note->user_id != Auth::id()) {
+            return $this->sendErorr('You do not have rights');
+        }
     	return view('notes.show')->with('note', $note);
     }
 
@@ -56,14 +69,17 @@ class NoteController extends Controller
      public function update(Request $request, $id)
     {
     	$note=Note::find($id);
-    	 $this->$validate($request,[
-
+    	 $input=$request->all();
+         $validator=Validator::make($input,[
             'title'=>'required',
             'content'=>'required'
         ]);
+    	  if ( $note->user_id != Auth::id()) {
+            return $this->sendErorr('You do not have rights');
+        }
 
-    	$note->title=$request->title;
-    	$note->content=$request->content;
+    	$note->title= $input['title'];
+        $note->content= $input['content'];
     	$note->save();
     	return redirect()->back();
 
@@ -71,9 +87,13 @@ class NoteController extends Controller
 
      public function destroy($id)
     {
-    	$note=Note::find($id);
-    	$note->delete();
-    	return redirect()->back();
+
+    	$note=Note::find($id); 
+    	 if ( $note->user_id != Auth::id()) {
+            return $this->sendErorr('You do not have rights');
+        }
+        $note->delete();
+        return redirect()->back();   	
     }
 
 }
